[FEAT] #38 - Product 리스트 부분은 product module로 분리
