+++ feature/product/src/main/java/com/peonlee/product/ProductViewModel.kt
@@ -1,4 +1,4 @@
-package com.peonlee.explore
+package com.peonlee.product
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -20,14 +20,13 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
 @HiltViewModel
-class ExploreViewModel @Inject constructor(
+class ProductViewModel @Inject constructor(
     private val productRepository: ProductRepository
 ) : ViewModel() {
     // 현재 상품 검색 조건
@@ -58,11 +57,7 @@ class ExploreViewModel @Inject constructor(
                     promotionRetailerList = productSearch.stores?.map { it.storeDataName },
                     promotionTypeList = productSearch.events?.map { it.eventDataName }
                 )
-            ).catch {
-                println(it)
-            }.map {
-                it.map { product -> product.toProductUiModel() }
-            }
+            ).map { it.map { product -> product.toProductUiModel() } }
         } catch (exception: Exception) {
             flowOf(PagingData.empty())
         }
