+++ feature/product/src/main/java/com/peonlee/product/adapter/FilterAdapter.kt
@@ -1,12 +1,12 @@
-package com.peonlee.explore.adapter
+package com.peonlee.product.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.peonlee.core.ui.adapter.SingleTypeListAdapter
 import com.peonlee.core.ui.viewholder.CommonViewHolder
-import com.peonlee.explore.adapter.viewholder.FilterViewHolder
-import com.peonlee.explore.databinding.ItemFilterSelectorBinding
-import com.peonlee.explore.model.BaseFilter
+import com.peonlee.product.adapter.viewholder.FilterViewHolder
+import com.peonlee.product.databinding.ItemFilterSelectorBinding
+import com.peonlee.product.model.BaseFilter
 
 class FilterAdapter(
     private val onClickEvent: (BaseFilter) -> Unit
