+++ settings.gradle.kts
@@ -49,3 +49,4 @@ include(":feature:detail")
 include(":core:model")
 include(":core:common")
 include(":feature:user")
+include(":feature:product")
